# swagger.yaml
swagger: "2.0"
info:
  title: Swidler API's documentation
  version: 1.0.0
...
paths:
...
  /answers:
    post:
      tags:
        - Answers
      summary: Create a new answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
      responses:
        200:
          description: Returns the created answer
        500:
          description: Server error

    get:
      tags:
        - Answers
      summary: Retrieve all answers
      responses:
        200:
          description: List of all answers
        500:
          description: Server error
  /answers/{id}:
    get:
      tags:
        - Answers
      summary: Retrieve a single answer by ID
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the answer to retrieve
          schema:
            type: integer
      responses:
        200:
          description: Returns the answer with the specified ID
        404:
          description: Answer not found
        500:
          description: Server error

    put:
      tags:
        - Answers
      summary: Update an answer by ID
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the answer to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
      responses:
        200:
          description: Returns the updated answer
        404:
          description: Answer not found
        500:
          description: Server error

    delete:
      tags:
        - Answers
      summary: Delete an answer by ID
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the answer to delete
          schema:
            type: integer
      responses:
        200:
          description: Answer deleted successfully
        404:
          description: Answer not found
        500:
          description: Server error
...
  /quiz/situations/{id}/answers:
    get:
      tags:
        - Quiz
      summary: Retrieve answers for a specific situation by ID
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the situation to retrieve answers for
          schema:
            type: integer
      responses:
        200:
          description: Returns the answers for the specified situation
        404:
          description: Answers not found for this situation
...
  /situations:
    post:
      tags:
        - Situations
      summary: Create a new situation
      consumes:
        - application/json
      parameters:
        - in: body
          name: situation
          description: The situation to create
          schema:
            type: object
            required:
              - question
            properties:
              question:
                type: string
              answerId:
                type: integer
              topicId:
                type: integer
              wrongAnswerDescription:
                type: string
      responses:
        200:
          description: Returns the created situation
        500:
          description: Server error

    get:
      tags:
        - Situations
      summary: Retrieve all situations
      produces:
        - application/json
      responses:
        200:
          description: List of all situations
        500:
          description: Server error
  /situations/{id}:
    get:
      tags:
        - Situations
      summary: Retrieve a single situation by ID
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: ID of the situation to retrieve
      responses:
        200:
          description: Returns the situation with the specified ID
        404:
          description: Situation not found
        500:
          description: Server error

    put:
      tags:
        - Situations
      summary: Update a situation by ID
      consumes:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: ID of the situation to update
        - in: body
          name: situation
          description: The situation details to update
          schema:
            type: object
            properties:
              question:
                type: string
              answerId:
                type: integer
              topicId:
                type: integer
              wrongAnswerDescription:
                type: string
      responses:
        200:
          description: Returns the updated situation
        404:
          description: Situation not found
        500:
          description: Server error

    delete:
      tags:
        - Situations
      summary: Delete a situation by ID
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: ID of the situation to delete
      responses:
        200:
          description: Situation deleted successfully
        404:
          description: Situation not found
        500:
          description: Server error